decouple test from specific entity
