+++ tests/datasets/validation/test_entity_validation.py
@@ -11,11 +11,7 @@ from snuba.query.data_source.simple import Entity as QueryEntity
 from snuba.query.exceptions import InvalidQueryException
 from snuba.query.expressions import Column, Expression, FunctionCall, Literal
 from snuba.query.logical import Query as LogicalQuery
-from snuba.query.validation.validators import (
-    ColumnValidationMode,
-    EntityContainsColumnsValidator,
-    EntityRequiredColumnValidator,
-)
+from snuba.query.validation.validators import EntityRequiredColumnValidator
 
 required_column_tests = [
     pytest.param(
@@ -124,44 +120,3 @@ def test_entity_required_column_validation_failure(
     validator = EntityRequiredColumnValidator(["project_id"])
     with pytest.raises(InvalidQueryException):
         validator.validate(query)
-
-
-entity_contains_columns_tests = [
-    pytest.param(
-        EntityKey.OUTCOMES,
-        id="Validate Outcomes Entity Columns",
-    )
-]
-
-
-@pytest.mark.parametrize("key", entity_contains_columns_tests)
-def test_outcomes_columns_validation(key: EntityKey) -> None:
-    entity = get_entity(key)
-
-    query_entity = QueryEntity(key, entity.get_data_model())
-
-    bad_query = LogicalQuery(
-        query_entity,
-        selected_columns=[
-            SelectedExpression("asdf", Column("_snuba_asdf", None, "asdf")),
-        ],
-    )
-
-    good_query = LogicalQuery(
-        query_entity,
-        selected_columns=[
-            SelectedExpression(
-                column.name, Column(f"_snuba_{column.name}", None, column.name)
-            )
-            for column in entity.get_data_model().columns
-        ],
-    )
-
-    validator = EntityContainsColumnsValidator(
-        entity.get_data_model(), validation_mode=ColumnValidationMode.ERROR
-    )
-
-    with pytest.raises(InvalidQueryException):
-        validator.validate(bad_query)
-
-    validator.validate(good_query)
